Initial commit to the bitbucket account. This consists of the latest version 2.0 software, completed at the end of summer 2013, and the more recent version 3.0 software, which was done by Matt Birkel in December 2013.
Version 3.0 includes a host-pc side simulation .vi, as well as the start
of framework to adjust wheel rotational speeds to allow for piviot points
within the robot's frame shadow on the ground.

File and folder names have also been updated to reflect the most recent
naming standards. See the technical wiki for this process, and please
continue to follow it for all future releases. Thanks!